successful zip retrieval from github and refactored package names from singular to plural
