--- src/main/java/com/revature/service/StorageService.java
@@ -1,8 +1,0 @@
-package com.revature.service;
-
-import org.springframework.web.multipart.MultipartFile;
-
-public interface StorageService {
-	void init();
-	String store(MultipartFile file);
-}
