+++ src/main/java/com/revature/services/FileService.java
@@ -1,0 +1,7 @@
+package com.revature.services;
+
+import org.springframework.http.ResponseEntity;
+
+public interface FileService {
+	ResponseEntity<byte[]> download(String fileURI);
+}
